% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructParents.R
\name{constructParents}
\alias{constructParents}
\title{Construct a dataframe with the parents}
\usage{
constructParents(x, keyvar, stub)
}
\arguments{
\item{x}{a dataframe with Parents variable}

\item{keyvar}{the primary key}

\item{stub}{Stub of variable name}
}
\value{
data
}
\description{
Construct a dataframe with the parents
}
