% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultings_ctu_report.R
\name{consultings_ctu_report}
\alias{consultings_ctu_report}
\title{Function to get the booked hours for consultings in total and in each year}
\usage{
consultings_ctu_report(all_tabs)
}
\arguments{
\item{all_tabs}{result from e.g. getPFData}
}
\value{
consultings
}
\description{
This code has been adapted for the purpose of producing the graphs for the QMR Yearly Report and CTU Annual Report.
The code gets (among others) the createYear of the ticket, customer (top_CustomerName, CustomerName, CustomerName2, Organization), default project of customer,
extracts whether it is a grant application consulting, runs prepTime to get total time bookings and time bookings for each year.
}
\examples{
all_tabs <-  getPFData()
x <- consultings_ctu_report(all_tabs)
}
