% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCustomFields.R
\name{decodeAllCustomFields}
\alias{decodeAllCustomFields}
\title{use the above decodeCustomFields function across all four instances of tables from pf
list = all_tabs
the direct backend download from projectfacts}
\usage{
decodeAllCustomFields(list)
}
\arguments{
\item{list}{}
}
\description{
use the above decodeCustomFields function across all four instances of tables from pf
list = all_tabs
the direct backend download from projectfacts
}
