% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCustomFields.R
\name{decodeCustomFields}
\alias{decodeCustomFields}
\title{Add custom fields to projects data}
\usage{
decodeCustomFields(
  dat,
  dat_ID_col,
  customfields,
  customfieldvalue,
  trace = FALSE
)
}
\arguments{
\item{customfields}{e.g. \code{all_tabs$customfields}}

\item{trace}{progress bar written to console, default = FALSE}

\item{project}{e.g. \code{all_tabs$project}}
}
\value{
data
}
\description{
Add custom fields to projects data
}
\examples{
all_tabs <- getPFData()
project <- all_tabs$project
customfields <- all_tabs$customfields
customfieldvalue <- all_tabs$customfieldvalue
project2 <- decodeCustomFields(project, customfields)
}
