% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_tdl_entries.R
\name{missing_tdl_entries}
\alias{missing_tdl_entries}
\title{Get missing TDL entries}
\usage{
missing_tdl_entries(all_tabs, rc_api_token)
}
\arguments{
\item{all_tabs}{pf data}

\item{rc_api_token}{API token to TDL on REDCap}
}
\value{
data
}
\description{
Get missing TDL entries
}
