% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_ctu_report.R
\name{projects_ctu_report}
\alias{projects_ctu_report}
\title{Function to get the booked hours for projects in total and in each year}
\usage{
projects_ctu_report(all_tabs)
}
\arguments{
\item{all_tabs}{result from e.g. getPFData}
}
\value{
projects
}
\description{
This code has been adapted for the purpose of producing the graphs for the QMR Yearly Report and CTU Annual Report. 
The code gets (among others) customer info (top_CustomerName, CustomerName, CustomerName2, Organization), default project of customer,
center info (mono/multi), ordinance info, archiving date of the project, runs prepTime to get total time bookings and time bookings for each year, 
categorizes sum of time bookings (< 8h, 9-20h,...).
}
\examples{
all_tabs <-  getPFData()
x <- projects_ctu_report(all_tabs)
}
