% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerActiveList.R
\name{workerActiveList}
\alias{workerActiveList}
\title{List of workers with their active status}
\usage{
workerActiveList(all_tabs)
}
\arguments{
\item{all_tabs}{e.g. the result from \code{getPFData()}}
}
\value{
data
}
\description{
List of workers with their active status
}
\examples{
all_tabs <- getPFData()
x <- setEncoding(all_tabs) 
y <- workerActiveList(x)

}
