% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCustomFields.R
\name{decodeCustomFields}
\alias{decodeCustomFields}
\title{Add custom fields to projects data}
\usage{
decodeCustomFields(dat, customfields, customfieldvalue)
}
\arguments{
\item{dat}{e.g. \code{all_tabs$project}}

\item{customfields}{e.g. \code{all_tabs$customfields}}

\item{customfieldvalue}{e.g. \code{all_tabs$customfieldvalue}}
}
\value{
data
}
\description{
Add custom fields to projects data
}
\examples{
all_tabs <- getPFData()
dat <- all_tabs$project
customfields <- all_tabs$customfields
customfieldvalue <- all_tabs$customfieldvalue
project2 <- decodeCustomFields(project, customfields, customfieldvalue)
}
