% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCustomFields.R
\name{decodeCustomFields}
\alias{decodeCustomFields}
\title{Add custom fields to projects data}
\usage{
decodeCustomFields(project, customfields, trace = FALSE)
}
\arguments{
\item{project}{e.g. \code{all_tabs$project}}

\item{customfields}{e.g. \code{all_tabs$customfields}}

\item{trace}{progress bar written to console, default = FALSE}
}
\value{
data
}
\description{
Add custom fields to projects data
}
\examples{
all_tabs <- getPFData()
project <- all_tabs$project
customfields <- all_tabs$customfields
project2 <- decodeCustomFields(project, customfields)
}
