% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sponsor.R
\name{get_sponsor}
\alias{get_sponsor}
\title{Function to get sponsoring institution and organization for projects/consultings}
\usage{
get_sponsor(input_data)
}
\arguments{
\item{input_data}{consultings/project data table with minimum variables CustomerName, CustomerName2, Organization, default_project}
}
\value{
output_data
}
\description{
This code has been adapted for the purpose of producing the graphs for the QMR Yearly Report and CTU Annual Report.
The code generates the variable SponsorOrganization (Insel, MedFac, UniBe Other, External Non/For Profit),
SponsorInstitution (e.g., Department at Inselspital), and SponsorInst_short (abbreviated institution-name to be used for plotting)
}
\examples{
all_tabs <-  getPFData()
input_data <-  consultings_ctu_report(all_tabs)
x <- get_sponsor(input_data)
}
